+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # OMC3 Changelog
 
+#### 2024-09-20 - v0.16.2 - _jdilly_ 
+
+- Added:
+  - Optics Measurements: `analyse_dpp` parameter, to analyse only files with a specific DPP for tune, phase and (C)RDTs.
+
 #### 2024-09-20 - v0.16.1 - _fsoubelet_
 
 - Fixed: